requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
pdfplumber
python-docx
spacy
spacy-transformers
faiss-cpu
numpy
ollama
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
scikit-learn
playwright